requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pytest
pytest-asyncio
httpx
ruff
mypy
pre-commit
